requirements.txt
beautifulsoup4==4.11.1
requests==2.27.1
lxml==4.9.1
pytz==2022.2.1
python-telegram-bot==13.7
